move header elements to new Header component
